requirements.txt
# --- Web Framework & Middleware ---
flask==2.2.5
flask-cors==3.0.10

# --- Deep Learning Model ---
torch==2.0.1
torchvision==0.15.2

# --- Image & Array Processing ---
pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72

# --- Augmentation (Model Preprocessing) ---
albumentations==1.3.1

# --- Visualization for Composite Summary ---
matplotlib==3.7.1

# --- WSGI Server for Production ---
gunicorn==20.1.0
